Split utl_ui project into UTL_BASE (non-GUI library) and UTL_UI (GUI) projects; extract CAppTools as abstract base for console vs GUI application objects

+++ IDETools/FileLocator.cpp
@@ -144,7 +144,7 @@ long FileLocator::GetSelectedCount( void )
 BSTR FileLocator::GetSelectedFile( long index )
 {
 	CString result;
-	if ( index >= 0 && index < m_selectedFiles.size() )
+	if ( index >= 0 && index < (long)m_selectedFiles.size() )
 		result = m_selectedFiles[ index ].first.GetPtr();
 	else
 		TRACE( _T("FileLocator::GetSelectedFile(): invalid index: %d from valid range [0, %d]\n"), index, m_selectedFiles.size() );
